requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.0
cffi==1.15.1
channels==4.0.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cryptography==42.0.8
daphne==4.0.0
dj-database-url==2.1.0
Django==3.2.23
django-cors-headers==4.1.0
django-extensions==3.2.3
django-redis==5.0.0
djangorestframework==3.12.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.7
importlib-metadata==4.8.3
incremental==22.10.0
kombu==5.1.0
packaging==23.2
Pillow==9.5.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.1.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.5.0
zipp==3.15.0
zope.interface==6.4.post2
